+++ client/src/views/contentTree/commands/buildAllGraphsAndTableListsCommand.ts
@@ -5,7 +5,6 @@ import * as vscode from 'vscode';
 import { SiemjConfigHelper } from '../../../models/siemj/siemjConfigHelper';
 import { SiemJOutputParser } from '../../../models/siemj/siemJOutputParser';
 import { Configuration } from '../../../models/configuration';
-import { SiemjConfBuilder } from '../../../models/siemj/siemjConfigBuilder';
 import { XpException } from '../../../models/xpException';
 import { DialogHelper } from '../../../helpers/dialogHelper';
 import { Log } from '../../../extension';
@@ -84,13 +83,15 @@ export class BuildAllGraphsAndTableListsCommand extends ViewCommand {
           } finally {
             const tmpPath = this.config.getTmpDirectoryPath(rootFolder);
             try {
-              // Очищаем временные файлы.
-              if (fs.lstatSync(tmpPath).isDirectory()) {
-                await fs.promises.rmdir(tmpPath, { recursive: true });
-              } else {
-                await fs.promises.access(tmpPath).then(() => {
-                  return fs.promises.unlink(tmpPath);
-                });
+              if (fs.existsSync(tmpPath)) {
+                // Очищаем временные файлы.
+                if (fs.lstatSync(tmpPath).isDirectory()) {
+                  await fs.promises.rmdir(tmpPath, { recursive: true });
+                } else {
+                  await fs.promises.access(tmpPath).then(() => {
+                    return fs.promises.unlink(tmpPath);
+                  });
+                }
               }
             } catch (e) {
               Log.warn('Clearing temporary files', e);
@@ -110,8 +111,8 @@ export class BuildAllGraphsAndTableListsCommand extends ViewCommand {
       if (!fs.existsSync(outputDirectory)) {
         fs.mkdirSync(outputDirectory, { recursive: true });
       }
-
-      const configBuilder = new SiemjConfBuilder(config, rootPath);
+      const siemjManager = new SiemjManager(this.config);
+      var configBuilder = siemjManager.getConfigBuilder(rootPath);
       configBuilder.addNormalizationsGraphBuilding();
       configBuilder.addAggregationGraphBuilding();
       configBuilder.addTablesSchemaBuilding();
